HaikuDepot: remove double space after full stop
Reported by translator loic (thanks!). The double space is only removed in
user-facing strings, I did not touch the comments.

For those interested in an internet controversy, search the internet for
the origins of the double space after a period. In any case, for consistency with
other user facing strings, go to a single space after a period.