StyledText translator: use a larger buffer for encoding detection
It may be the case that a file is largely ASCII with few UTF-8 or
similar characters (typically, english text mentionning names in other
languages). It turns out 2 kilobytes of text is not enough for a safe
guess in that case.

Fixes #14674

Change-Id: Id894bebb821c22a17b776a71822d011abc795dea
Reviewed-on: https://review.haiku-os.org/c/haiku/+/2628
Reviewed-by: waddlesplash <waddlesplash@gmail.com>
(cherry picked from commit 8bb5fd243743acca90bc9fc0a3c03abee5f674f5)
Reviewed-on: https://review.haiku-os.org/c/haiku/+/2636