usb_hid: do not publish a keyboard device when a mouse is connected
My trackball has a "system control" in its descriptor, but is
definitely not a keyboard. This results in an extra entry in Input
preferences for it.

If we actually want to accept input from such devices, we should scan
them more closely to check there is actually a valid input there, and
consider publishing them as something else than keyboards, because
that's not really suitable in Input preferences.

Change-Id: I9027454b3b24d0f39ea562aed851402d0be7aeb4
Reviewed-on: https://review.haiku-os.org/c/haiku/+/2715
Reviewed-by: waddlesplash <waddlesplash@gmail.com>