Beta2: block using the Installer to upgrade an existing system.
See #15808. Installer does not do what it promises to do, that is, that is
clean the system folder and essentially perform a clean install. Instead all
packages are copied to /system/packages, but they are not activated. To make it
worse, the system is then unable to upgrade through pkgman later on, as it
bails with the error that the files already exist.

This change at least stops Installer from pretending to do a clean install.

Change-Id: I5e2b1f571a99f9f8c020f77c314974719346f750
Reviewed-on: https://review.haiku-os.org/c/haiku/+/2760
Reviewed-by: Adrien Destugues <pulkomandy@gmail.com>