kernel/thread: Fix team double lock when reaching thread limit.
The team is locked for the allocation of the user thread. When the
thread limit is reached, the user thread is removed again by
team_free_user_thread which internally locks the team, resulting in a
double lock. Unlock the teamLocker before calling team_free_user_thread.

Amazingly noone seems to ever have hit the thread limit, even though it
is currently at an arbitrarily low 4096.